move cg into optimizers
