+++ src/mrpro/algorithms/optimizers/__init__.py
@@ -1,2 +1,3 @@
 from mrpro.algorithms.optimizers._adam import adam
+from mrpro.algorithms.optimizers._cg import cg
 from mrpro.algorithms.optimizers._lbfgs import lbfgs
